feat!(connection): slim down `IpmiCommand` trait.
Update the `IpmiCommand` trait to reduce/remove completion-code validation
responsibility from implementors of that trait.

The trait's functions are updated to reflect the two distrinct use-cases: special handling
of failure completion codes for specific commands, and parsing the data provided
by a message with a success completion code.

This also allows for removal of a superfluous layer of errors from the `Ipmi` struct.

Add a specific `NotEnoughData` unit-type for not enough data.