fix(ipmi-rs/examples): don't panic when a sensor is not present
Sometimes, a sensor that exits in the SDR is not present. This may occur
when e.g. a fan header or NVMe slot is present on the motherboard but
not populated on a system. Currently, the `get-info` example panics when
encountering a sensor that isn't present, which is a little unfortunate.

This commit changes the example to just log that the sensor is not
present and keep going. When there are other errors, we also just log
the error (at the `Error` level) and keep going, so that the user can
still see other SDRs that are read successfully.

Fixes datdenkikniet#ipmi-rs#34